Create and populate `plasmapy.formulary.misc` (#1453)
Excellent! Thank you.
