+++ changelog/1453.feature.rst
@@ -1,0 +1,10 @@
+Created `plasmapy.formulary.misc` to contain functionality for
+miscellaneous plasma parameters, and migrated
+``~plasmapy.formulary.misc._grab_charge``,
+`~plasmapy.formulary.misc.Bohm_diffusion`,
+`~plasmapy.formulary.misc.magnetic_energy_density`,
+`~plasmapy.formulary.misc.magnetic_pressure`,
+`~plasmapy.formulary.misc.mass_density`, and
+`~plasmapy.formulary.misc.thermal_pressure` from
+`plasmapy.formulary.parameters` to the new module.  Related aliases were
+also migrated.
