+++ changelog/1453.removal.rst
@@ -1,0 +1,2 @@
+Officially deprecated `plasmapy.formulary.parameters` and scheduled its
+permanent removal for the ``v0.9.0`` release.
