+++ docs/formulary/misc.rst
@@ -1,0 +1,7 @@
+.. _formulary_misc:
+
+***********************************************************
+Miscellaneous Plasma Parameters (`plasmapy.formulary.misc`)
+***********************************************************
+
+.. automodapi:: plasmapy.formulary.misc
