+++ docs/formulary/speeds.rst
@@ -1,0 +1,13 @@
+.. _speeds:
+
+*****************************************************
+Speed Plasma Parameters (`plasmapy.formulary.speeds`)
+*****************************************************
+
+.. automodapi:: plasmapy.formulary.speeds
+
+.. nbgallery::
+    :caption: Examples
+
+    /notebooks/distribution
+    /notebooks/formulary/thermal_speed
