+++ plasmapy/formulary/misc.py
@@ -1,0 +1,413 @@
+"""Functions for miscellaneous plasma parameter calculations."""
+
+__all__ = [
+    "Bohm_diffusion",
+    "magnetic_energy_density",
+    "magnetic_pressure",
+    "mass_density",
+    "thermal_pressure",
+]
+__aliases__ = ["DB_", "pmag_", "pth_", "rho_", "ub_"]
+
+import astropy.units as u
+import numbers
+import numpy as np
+
+from astropy.constants.si import e, k_B, mu0
+from typing import Optional, Union
+
+from plasmapy import particles
+from plasmapy.particles import Particle
+from plasmapy.utils.decorators import validate_quantities
+
+
+def _grab_charge(ion: Particle, z_mean=None):
+    """
+    Merge two possible inputs for particle charge.
+
+    Parameters
+    ----------
+    ion : `~plasmapy.particles.particle_class.Particle`
+        a string representing a charged particle, or a Particle object.
+
+    z_mean : `float`
+        An optional float describing the average ionization of a particle
+        species.
+
+    Returns
+    -------
+    float
+        if ``z_mean`` was passed, ``z_mean``, otherwise, the charge number
+        of ``ion``.
+
+    """
+    if z_mean is None:
+        # warnings.warn("No z_mean given, defaulting to atomic charge",
+        #               PhysicsWarning)
+        Z = particles.charge_number(ion)
+    else:
+        # using average ionization provided by user
+        Z = z_mean
+    return Z
+
+
+@validate_quantities(
+    T_e={"can_be_negative": False, "equivalencies": u.temperature_energy()},
+    B={"can_be_negative": False},
+)
+def Bohm_diffusion(T_e: u.K, B: u.T) -> u.m ** 2 / u.s:
+    r"""
+    Return the Bohm diffusion coefficient.
+
+    The Bohm diffusion coefficient was conjectured to follow Bohm model
+    of the diffusion of plasma across a magnetic field and describe the
+    diffusion of early fusion energy machines :cite:p:`bohm:1949`. The
+    rate predicted by Bohm diffusion is much higher than classical
+    diffusion, and if there were no exceptions, magnetically confined
+    fusion would be impractical.
+
+    .. math::
+
+        D_B = \frac{1}{16} \frac{k_B T}{e B}
+
+    where :math:`k_B` is the Boltzmann constant
+    and :math:`e` is the fundamental charge.
+
+    **Aliases:** `DB_`
+
+    Parameters
+    ----------
+    T_e : `~astropy.units.Quantity`
+        The electron temperature.
+
+    B : `~astropy.units.Quantity`
+        The magnitude of the magnetic field in the plasma.
+
+    Warns
+    -----
+    : `~astropy.units.UnitsWarning`
+        If units are not provided, SI units are assumed.
+
+    Raises
+    ------
+    `TypeError`
+        ``T_e`` is not a `~astropy.units.Quantity` and cannot be
+        converted into one.
+
+    `~astropy.units.UnitConversionError`
+        If ``T_e`` is not in appropriate units.
+
+    Examples
+    --------
+    >>> import astropy.units as u
+    >>> T_e = 5000 * u.K
+    >>> B = 10 * u.T
+    >>> Bohm_diffusion(T_e, B)
+    <Quantity 0.00269292 m2 / s>
+    >>> T_e = 50 * u.eV
+    >>> B = 10 * u.T
+    >>> Bohm_diffusion(T_e, B)
+    <Quantity 0.3125 m2 / s>
+
+    Returns
+    -------
+    D_B : `~astropy.units.Quantity`
+        The Bohm diffusion coefficient in meters squared per second.
+
+    """
+    D_B = k_B * T_e / (16 * e * B)
+    return D_B
+
+
+DB_ = Bohm_diffusion
+"""Alias to `~plasmapy.formulary.misc.Bohm_diffusion`."""
+
+
+@validate_quantities
+def magnetic_energy_density(B: u.T) -> u.J / u.m ** 3:
+    r"""
+    Calculate the magnetic energy density.
+
+    **Aliases:** `ub_`
+
+    Parameters
+    ----------
+    B : `~astropy.units.Quantity`
+        The magnetic field in units convertible to tesla.
+
+    Returns
+    -------
+    E_B : `~astropy.units.Quantity`
+        The magnetic energy density in units of joules per cubic meter.
+
+    Raises
+    ------
+    `TypeError`
+        If the input is not a `~astropy.units.Quantity`.
+
+    `~astropy.units.UnitConversionError`
+        If the input is not in units convertible to tesla.
+
+    `ValueError`
+        If the magnetic field strength does not have an appropriate
+        value.
+
+    Warns
+    -----
+    : `~astropy.units.UnitsWarning`
+        If units are not provided, SI units are assumed
+
+    Notes
+    -----
+    The magnetic energy density is given by:
+
+    .. math::
+        E_B = \frac{B^2}{2 μ_0}
+
+    The motivation behind having two separate functions for magnetic
+    pressure and magnetic energy density is that it allows greater
+    insight into the physics that are being considered by the user and
+    thus more readable code.
+
+    See Also
+    --------
+    magnetic_pressure : Returns an equivalent `~astropy.units.Quantity`,
+        except in units of pascals.
+
+    Examples
+    --------
+    >>> from astropy import units as u
+    >>> magnetic_energy_density(0.1*u.T)
+    <Quantity 3978.87... J / m3>
+
+    """
+    return magnetic_pressure(B)
+
+
+ub_ = magnetic_energy_density
+"""Alias to `~plasmapy.formulary.misc.magnetic_energy_density`."""
+
+
+@validate_quantities
+def magnetic_pressure(B: u.T) -> u.Pa:
+    r"""
+    Calculate the magnetic pressure.
+
+    **Aliases:** `pmag_`
+
+    Parameters
+    ----------
+    B : `~astropy.units.Quantity`
+        The magnetic field in units convertible to tesla.
+
+    Returns
+    -------
+    p_B : `~astropy.units.Quantity`
+        The magnetic pressure in units in pascals (newtons per square meter).
+
+    Raises
+    ------
+    `TypeError`
+        If the input is not a `~astropy.units.Quantity`.
+
+    `~astropy.units.UnitConversionError`
+        If the input is not in units convertible to tesla.
+
+    `ValueError`
+        If the magnetic field strength is not a real number between
+        :math:`±∞`\ .
+
+    Warns
+    -----
+    : `~astropy.units.UnitsWarning`
+        If units are not provided, SI units are assumed.
+
+    Notes
+    -----
+    The magnetic pressure is given by:
+
+    .. math::
+        p_B = \frac{B^2}{2 μ_0}
+
+    The motivation behind having two separate functions for magnetic
+    pressure and magnetic energy density is that it allows greater
+    insight into the physics that are being considered by the user and
+    thus more readable code.
+
+    See Also
+    --------
+    magnetic_energy_density : returns an equivalent `~astropy.units.Quantity`,
+        except in units of joules per cubic meter.
+
+    Examples
+    --------
+    >>> from astropy import units as u
+    >>> magnetic_pressure(0.1*u.T).to(u.Pa)
+    <Quantity 3978.87... Pa>
+
+    """
+    return (B ** 2) / (2 * mu0)
+
+
+pmag_ = magnetic_pressure
+"""Alias to `~plasmapy.formulary.misc.magnetic_pressure`."""
+
+
+@validate_quantities(
+    density={"can_be_negative": False}, validations_on_return={"can_be_negative": False}
+)
+def mass_density(
+    density: (u.m ** -3, u.kg / (u.m ** 3)),
+    particle: Union[Particle, str],
+    z_ratio: Optional[numbers.Real] = 1,
+) -> u.kg / u.m ** 3:
+    r"""
+    Calculate the mass density from a number density.
+
+    .. math::
+
+        \rho = \left| \frac{Z_{s}}{Z_{particle}} \right| n_{s} m_{particle}
+              = | Z_{ratio} | n_{s} m_{particle}
+
+    where :math:`m_{particle}` is the particle mass, :math:`n_{s}` is a number
+    density for plasma species :math:`s`, :math:`Z_{s}` is the charge number of
+    species :math:`s`, and :math:`Z_{particle}` is the charge number of
+    ``particle``.  For example, if the electron density is given for :math:`n_s`
+    and ``particle`` is a doubly ionized atom, then :math:`Z_{ratio} = -1 / 2`\ .
+
+    **Aliases:** `rho_`
+
+    Parameters
+    ----------
+    density : `~astropy.units.Quantity`
+        Either a particle number density (in units of m\ :sup:`-3` or
+        equivalent) or a mass density (in units of kg / m\ :sup:`3` or
+        equivalent).  If ``density`` is a mass density, then it will be passed
+        through and returned without modification.
+
+    particle : `~plasmapy.particles.particle_class.Particle`
+        The particle for which the mass density is being calculated for.  Must
+        be a `~plasmapy.particles.particle_class.Particle` or a value convertible to
+        a `~plasmapy.particles.particle_class.Particle` (e.g., ``'p'`` for protons,
+        ``'D+'`` for deuterium, or ``'He-4 +1'`` for singly ionized helium-4).
+
+    z_ratio : `int`, `float`, optional
+        The ratio of the charge numbers corresponding to the plasma species
+        represented by ``density`` and the ``particle``.  For example, if the
+        given ``density`` is and electron density and ``particle`` is doubly
+        ionized ``He``, then ``z_ratio = -0.5``.  Default is ``1``.
+
+    Raises
+    ------
+    `~astropy.units.UnitTypeError`
+        If the ``density`` does not have units equivalent to a number density
+        or mass density.
+
+    `TypeError`
+        If ``density`` is not of type `~astropy.units.Quantity`, or convertible.
+
+    `TypeError`
+        If ``particle`` is not of type or convertible to
+        `~plasmapy.particles.particle_class.Particle`.
+
+    `TypeError`
+        If ``z_ratio`` is not of type `int` or `float`.
+
+    `ValueError`
+        If ``density`` is negative.
+
+    Returns
+    -------
+    `~astropy.units.Quantity`
+        The mass density for the plasma species represented by ``particle``.
+
+    Examples
+    --------
+    >>> import astropy.units as u
+    >>> mass_density(1 * u.m ** -3, 'p')
+    <Quantity 1.67262...e-27 kg / m3>
+    >>> mass_density(4 * u.m ** -3, 'D+')
+    <Quantity 1.33743...e-26 kg / m3>
+    >>> mass_density(2.e12 * u.cm ** -3, 'He')
+    <Quantity 1.32929...e-08 kg / m3>
+    >>> mass_density(2.e12 * u.cm ** -3, 'He', z_ratio=0.5)
+    <Quantity 6.64647...e-09 kg / m3>
+    >>> mass_density(1.0 * u.g * u.m ** -3, "")
+    <Quantity 0.001 kg / m3>
+    """
+    if density.unit.is_equivalent(u.kg / u.m ** 3):
+        return density
+
+    if not isinstance(particle, Particle):
+        try:
+            particle = Particle(particle)
+        except TypeError:
+            raise TypeError(
+                f"If passing a number density, you must pass a plasmapy Particle "
+                f"(not type {type(particle)}) to calculate the mass density!"
+            )
+
+    if not isinstance(z_ratio, (float, np.floating, int, np.integer)):
+        raise TypeError(
+            f"Expected type int or float for keyword z_ratio, got type {type(z_ratio)}."
+        )
+
+    return abs(z_ratio) * density * particle.mass
+
+
+rho_ = mass_density
+"""Alias to `~plasmapy.formulary.misc.mass_density`."""
+
+
+@validate_quantities(
+    T={"can_be_negative": False, "equivalencies": u.temperature_energy()},
+    n={"can_be_negative": False},
+)
+def thermal_pressure(T: u.K, n: u.m ** -3) -> u.Pa:
+    r"""
+    Return the thermal pressure for a Maxwellian distribution.
+
+    **Aliases:** `pth_`
+
+    Parameters
+    ----------
+    T : `~astropy.units.Quantity`
+        The particle temperature in either kelvin or energy per particle.
+
+    n : `~astropy.units.Quantity`
+        The particle number density in units convertible to m\ :sup:`-3`\ .
+
+    Examples
+    --------
+    >>> import astropy.units as u
+    >>> thermal_pressure(1*u.eV, 1e20/u.m**3)
+    <Quantity 16.021... Pa>
+    >>> thermal_pressure(10*u.eV, 1e20/u.m**3)
+    <Quantity 160.21... Pa>
+
+    Returns
+    -------
+    p_th : `~astropy.units.Quantity`
+        Thermal pressure.
+
+    Raises
+    ------
+    `TypeError`
+        The temperature or number density is not a `~astropy.units.Quantity`.
+
+    `~astropy.units.UnitConversionError`
+        If the particle temperature is not in units of temperature or
+        energy per particle.
+
+    Notes
+    -----
+    The thermal pressure is given by:
+
+    .. math::
+        T_{th} = n k_B T
+    """
+    return n * k_B * T
+
+
+pth_ = thermal_pressure
+"""Alias to `~plasmapy.formulary.misc.thermal_pressure`."""
